fix bad why3 term generation
Binary operations were constructed using Tbinop which is used for
chaining, leading to a term such as "(a -> b) && (b -> c)" printed as
"a -> b && b -> c", which is then parsed as "(a -> b && b) -> c" by Why3.

Solution is to use the Tbinnop constructor which does not allow chaining
and prints as expected.