don't put transition eba in CNF
this creates additional predicates to stay linear in the size of the input formula which obscure what the user must prove and
requires additional support from the backend: every predicate introduced that is not in the precondition but present in the postcondition must be allowed to be assigned any truth value. Otherwise, the goal cannot be solved.